pyproject.toml
[tool.poetry]
name = "masak2"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
psycopg2 = "^2.8.5"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.15.0"
pytest-django = "^3.9.0"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.vscode
    | _build
    | buck-out
    | build
    | dist
    | docker

  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''